Added Carrier OPS
CARRIER OPERATIONS REFERENCE CARD

After realizing there's no quick reference available for the comms and pattern work involved in CAW carrier ops, I decided to start work on a CAW carrier ops reference card.

The first part (communications calls) are derived straight from the CAW CASE1 Manual, formatted, abbreviated and color-coded for easy reference.

The second part (pattern and carrier diagram) I wasn't originally going to include, but realized I have juuuust enough room after prepping the comms section to go ahead and try to do it anyways. Because it wasn't formatted in landscape originally, I had to completely re-do the text from scratch, which came with the benefit of being able to color code the diagram text to the corresponding pattern sections to make it (imo) much easier to read.

This is not intended to teach you carrier ops - go read through the CAW CASE1 manual for that. Instead, this is intended to be used by someone who has already gone through the manual for quick reference to help smooth out the workflow during ops and to provide a quick refresher for someone who's been away for a bit.

The full source code and files are available on my github, as always :D